Processing mode (#19)
* Image processing mode

* log

* delete-all entities

* readme update

+++ Api/Api/Services/GraphService.cs
@@ -85,7 +85,8 @@ namespace Api.Services
             var image = GetGraphFile(guid, GraphFormat.RawImage).File;
 
             var script = _appSettings.StoragePaths.ScriptFullPath;
-            var param = "-p " + image.FullName;// + " -b " + (int) mode;
+            var param = "-p " + image.FullName + " -b " + model.Mode.ToString();
+            _logger.LogInformation("SCRIPT PARAMETERS: " + param);
 
             var processResult = new PythonRunner().Run(script, param);
             _logger.LogInformation("Processing image result:\n " + processResult);
@@ -153,6 +154,25 @@ namespace Api.Services
                 _dataContext.GraphEntities.Remove(entity);
 
                 var path = Path.Combine(_appSettings.StoragePaths.UploadsDirectory, guid.ToString());
+                if (Directory.Exists(path))
+                    Directory.Delete(path, true);
+            }
+
+            await _dataContext.SaveChangesAsync();
+
+            return true;
+        }
+
+        public async Task<bool> RemoveAllUsersEntitiesAsync(int userId)
+        {
+            var user = _dataContext.Users.Include(u=>u.GraphEntities).SingleOrDefault(u => u.Id == userId);
+            if (user == null) return false;
+
+            foreach (var entity in user.GraphEntities)
+            {
+                _dataContext.GraphEntities.Remove(entity);
+
+                var path = Path.Combine(_appSettings.StoragePaths.UploadsDirectory, entity.GUID.ToString());
                 if (Directory.Exists(path))
                     Directory.Delete(path, true);
             }
